Cleaned Fragments, Amended MainScroll, Produced Placeholder Fragment Classes
I’ve cleaned some superfluous code from the pre-amble fragments, now I
just have to implement a simpler carousel and we’re flying.

The main scroll now send a message to the section it creates when you
click on a body part, and the section itself populates itself currently
with the correct icon, header and interactive fragment (the ones you
guys are working on!).

I’ve made some placeholder classes and related XML for you to place
your code into.